feat: 계좌 생성, 삭제, 입금, 출금 기능 구현

+++ src/main/java/zerobase/fintech/controller/AccountController.java
@@ -6,10 +6,9 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,9 +17,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import zerobase.fintech.dto.AccountDto;
-import zerobase.fintech.dto.DeleteAccountDto;
-import zerobase.fintech.dto.DepositWithdrawDto;
+import zerobase.fintech.dto.account.AccountDto;
+import zerobase.fintech.dto.account.DeleteAccountDto;
+import zerobase.fintech.dto.account.DepositWithdrawDto;
 import zerobase.fintech.entity.Account;
 import zerobase.fintech.service.AccountService;
 
@@ -28,16 +27,24 @@ import zerobase.fintech.service.AccountService;
 @RequiredArgsConstructor
 @RequestMapping("/account")
 public class AccountController {
+
   private final AccountService accountService;
+  private final MemberController memberController;
 
   /**
    * 계좌 생성 하는 API
+   *
    * @param accountDto
    * @param email
    * @return
    */
-  @PostMapping("/create/{email}")
-  public ResponseEntity<?> createAccount(@RequestBody AccountDto accountDto, @PathVariable String email){
+  @PostMapping("/{email}")
+  public ResponseEntity<?> createAccount(@Validated @RequestBody AccountDto accountDto, BindingResult bindingResult, @PathVariable String email) {
+    ResponseEntity<?> response = memberController.getErrorResponseEntity(bindingResult);
+    if (response != null) {
+      return response;
+    }
+
     Account account = accountService.createAccount(email, accountDto);
     LinkedHashMap<String, Object> result = new LinkedHashMap<>();
     result.put("status", "계좌 생성 성공");
@@ -49,13 +56,19 @@ public class AccountController {
 
   /**
    * 계좌 조회하는 API
+   *
    * @param accountDto
    * @param email
    * @param page
    * @return
    */
-  @GetMapping("/find/{email}")
-  public ResponseEntity<?> selectAccount(@RequestBody AccountDto accountDto, @PathVariable String email, @RequestParam(value="page", defaultValue="0") int page){
+  @GetMapping("/{email}")
+  public ResponseEntity<?> selectAccount(@Validated @RequestBody AccountDto accountDto, BindingResult bindingResult, @PathVariable String email, @RequestParam(value = "page", defaultValue = "0") int page) {
+    ResponseEntity<?> response = memberController.getErrorResponseEntity(bindingResult);
+    if (response != null) {
+      return response;
+    }
+
     List<Account> accountList = accountService.findAccount(email, accountDto, page);
     List<Map<String, Object>> accountResult = new ArrayList<>();
 
@@ -74,12 +87,19 @@ public class AccountController {
 
   /**
    * 계좌 삭제 하는 API
+   *
    * @param deleteAccountDto
    * @param email
    * @return
    */
-  @DeleteMapping("/delete/{email}")
-  public ResponseEntity<?> deleteAccount(@RequestBody DeleteAccountDto deleteAccountDto, @PathVariable String email){
+  @DeleteMapping("/{email}")
+  public ResponseEntity<?> deleteAccount(@Validated @RequestBody DeleteAccountDto deleteAccountDto,
+      BindingResult bindingResult, @PathVariable String email) {
+    ResponseEntity<?> response = memberController.getErrorResponseEntity(bindingResult);
+    if (response != null) {
+      return response;
+    }
+
     String accountNum = accountService.deleteAccount(email, deleteAccountDto);
 
     LinkedHashMap<String, Object> result = new LinkedHashMap<>();
@@ -92,12 +112,18 @@ public class AccountController {
 
   /**
    * 계좌 입금 하는 API
+   *
    * @param depositWithdrawDto
    * @param accountNum
    * @return
    */
   @PostMapping("/deposit/{accountNum}")
-  public ResponseEntity<?> depositAccount(@RequestBody DepositWithdrawDto depositWithdrawDto, @PathVariable String accountNum){
+  public ResponseEntity<?> depositAccount(@Validated @RequestBody DepositWithdrawDto depositWithdrawDto, BindingResult bindingResult, @PathVariable String accountNum) {
+    ResponseEntity<?> response = memberController.getErrorResponseEntity(bindingResult);
+    if (response != null) {
+      return response;
+    }
+
     Account account = accountService.depositAccount(accountNum, depositWithdrawDto);
 
     LinkedHashMap<String, Object> result = new LinkedHashMap<>();
@@ -109,12 +135,18 @@ public class AccountController {
 
   /**
    * 계좌 출금 하는 API
+   *
    * @param depositWithdrawDto
    * @param accountNum
    * @return
    */
   @PostMapping("/withdraw/{accountNum}")
-  public ResponseEntity<?> withdrawAccount(@RequestBody DepositWithdrawDto depositWithdrawDto, @PathVariable String accountNum){
+  public ResponseEntity<?> withdrawAccount(@Validated @RequestBody DepositWithdrawDto depositWithdrawDto, BindingResult bindingResult, @PathVariable String accountNum) {
+    ResponseEntity<?> response = memberController.getErrorResponseEntity(bindingResult);
+    if (response != null) {
+      return response;
+    }
+
     Account account = accountService.withdrawAccount(accountNum, depositWithdrawDto);
 
     LinkedHashMap<String, Object> result = new LinkedHashMap<>();
